Move event to models directory
