+++ app/app.js
@@ -16,7 +16,7 @@ import {
   versionHandler,
 } from './handlers/index.js';
 import Context from './context.js';
-import Event from './event.js';
+import Event from './models/event.js';
 
 /**
  * @param {Context} context
